separated library and cli
